feat: modularize globe showcase component by extracting globe container and animation logic
